Adapt to new commons dependency moves and migrated ExecutionGraph here as this is the only module using it
